Refactoring the Clients
I wanted to simplify the code that will be user facing by refactoring both the UserClient and ChallengeClient into a unified Client class.

Furthermore I have removed the requirement for an HTTPClient instance to be passed into the constructor of the new Client. I feel that this is a step most users will find cumbersome and to this end, the codebase now uses GuzzleHttp under the hood and tries to maintain some semblance of tidyness but more refactoring in this area will continue as v3 develops.